Handwritten-Eqn-Solver-main/HandCalc/requirements.txt
keras==2.9.0
ipykernel
opencv-python
numpy==1.26.4
pandas
pillow
sympy
matplotlib